requirements.txt
opencv-python
pillow
python-vlc
